+++ packages/core/src/extensions/HyperlinkToolbar/HyperlinkToolbarPlugin.ts
@@ -6,7 +6,7 @@ import { BlockNoteEditor } from "../../editor/BlockNoteEditor";
 import { BlockSchema } from "../../schema/blocks/types";
 import { InlineContentSchema } from "../../schema/inlineContent/types";
 import { StyleSchema } from "../../schema/styles/types";
-import { BaseUiElementState } from "../@util/BaseUiElementTypes";
+import { BaseUiElementState } from "../@shared/BaseUiElementTypes";
 import { EventEmitter } from "../../@util/EventEmitter";
 
 export type HyperlinkToolbarState = BaseUiElementState & {
